requirements.txt
streamlit
tensorflow
ultralytics
numpy
pillow
